Remove thor/rake from Gemfile
This is mostly a slight convenience thing than anything.

For obvious reasons, `thor` is removed since it is not a direct
dependency.

Removing `rake` is a little more nuanced.  When trying to run `rake` for
anything outside of `rake -T`, the `bundler/setup` will require most
likely cause an error since it will try loading the version of rake in
the Gemfile instead of the one that most likely the one that didn't
start the process (the one included with Ruby).

Since Ruby already ships with a rake, let's just use that one, since the
older versions should be completely compatible with anything we are
doing, and has been in Ruby since at least the 1.9 days... which is way
older than any Ruby version we support with this project.